Change ID scalar from string to interface{}.
The ID is documented as an opaque type. The user should not be looking
at its contents or modifying it:

	The ID type appears in a JSON response as a String; however, it is
	not intended to be human-readable. When expected as an input type,
	any string (such as "VXNlci0xMA==") or integer (such as 4) input
	value will be accepted as an ID.

So, it seems that using a more opaque interface{} underlying type is in
spirit of the ID type.

This is an experimental change. We will try it and see how it goes, and
revert if it turns out to be a bad idea.