Change type of URI and X509Certificate scalars to be more structured.
The idea is similar to DateTime type, it might be helpful to further
parse the URI into url.URL and X509Certificate into x509.Certificate.
That way the users don't have to do it ad-hoc.

If the user prefers to deal with string-encoded versions of these
types, they can use string type in the query.

json.Marshaler methods for X509Certificate type are not implemented,
because I don't have access to any sample values for the GitHub
response at this time. They will be implemented later.

This is an experimental change. We will try it and see how it goes, and
revert if it turns out to be a bad idea.