Add input parameter to Client.Mutate method.
This is a breaking API change.

Every mutation has some input fields, so they'll require a corresponding
input object. Make it more convenient to pass the input, especially when
there are no other variables that are needed.

This change is strictly for convenience, because it seems like it leads
to a cleaner API. Time and usage will tell whether this is a good change
or if it's unhelpful and should be reverted.