README: Simplify first variables example.
It makes sense to give variables a longer, more descriptive name when
there are lots of them. But in this simple example, there are only two,
so the "Repository" prefix is unnecessary. It can also be confusing.

It's already present in the next example on pagination, where it makes
sense because there are more variables involved. It helps to gradually
give more advanced examples and build on the lessons from previous ones.

Thanks to @slimsag for helping me spot this opportunity.