Rename variables to begin with lowercase.
This is more idiomatic GraphQL style. There used to be a technical
reason that they had to begin with an uppercase letter before #4,
but that's no longer the case.

Since there's no reason to deviate from GraphQL style here, the plan
is to follow it more closely for now. That is, until there's a good
reason to deviate from it.

This change is really just a documentation/example change. Users can
still name variables whatever they want.

Resolves #5.