Resolve name conflicts in generated enum value identifiers.
Go with solution 1, as described in #8. Solution 1 is the simplest,
most predictable, and least controversial option. Its only downfall
are the verbose enum value identifiers, but it suffers from least
other problems.

Since the API of the library is not frozen yet, if another solution
is found which is determined to be better overall, we can still
improve this.

Regenerate with:

	go generate ./...

Updates #8.
Follows #7.