Assume that ID has string underlying type.
This workaround is possible since ID is the only type that can result
in the underlying type being string. For normal strings,
githubql.String type would be used.

This change adds a workaround and a passing test case. If a better
solution can be found, it can be applied as long as tests still pass.

Helps #12.