Rename package from githubql to githubv4.
This should be a more appropriate and less confusing name.
See issue #35 for full rationale.

This is a breaking API change. Updating code is easy.
It can be done with an automated tool or by hand.

The gomvpkg tool (https://golang.org/x/tools/cmd/gomvpkg)
can be used to rename your existing githubql package to
githubv4, and update all code references in your GOPATH:

	gomvpkg -from github.com/shurcooL/githubql -to github.com/shurcooL/githubv4

(You'll need to run it before you update to this commit.)

Otherwise, if you want to update your code manually,
only the import declaration needs to be changed at first:

	 import (
	 	// ...
	-	"github.com/shurcooL/githubql"
	+	githubql "github.com/shurcooL/githubv4"
	 )

Afterwards, it's possible to update code to use the new
githubv4 package name and remove the import rename.

Or, if you've already renamed the package to githubv4:

	 import (
	 	// ...
	 	githubv3 "github.com/google/go-github/github"
	-	githubv4 "github.com/shurcooL/githubql"
	+	"github.com/shurcooL/githubv4"
	 )

Resolves #35.