regenerate for schema changes by 2019-06-01
Start cleaning descriptions by using new "clean" template function.

Some of the GraphQL descriptions started being split across multiple
lines via newline characters. That was not compatible with how we're
generating the line comments, because the assumption was that the
description was a single line.

For now, it's viable to clean them by replacing all whitespace
characters with a single space, effectively inlining their text.
This works because none of the descriptions we're using at this
time contain multiple paragraphs. Should that change, we'll need
to start using a different strategy for generating comments.

Regenerated with:

	GITHUB_TOKEN=<redacted> go generate ./...

Reference: https://developer.github.com/v4/changelog/.