regenerate for schema changes by 2023-02-14
Done with:

	GITHUB_TOKEN=<redacted> go generate ./...

Add special handling of the SCIM enum value to gen.go to generate a
better identifier name for it (per https://go.dev/s/style#initialisms).
In gen.go for now because it's easier than teaching ident about it,
and because it's not very common (at this point).

Reference: https://docs.github.com/en/graphql/overview/changelog.